fix(ml[by oscar]): the load_model issue with memory leak has been fixed
